split out instantiation tests
