+++ tests/parser/exceptions/test_invalid_type_exception.py
@@ -72,97 +72,6 @@ x: Bytes <= wei
     """
 x: 5
     """,
-    """
-event Foo:
-    a: uint256
-
-@external
-def foo() -> Foo:
-    return Foo(2)
-    """,
-    """
-event Foo:
-    a: uint256
-
-@external
-def foo() -> (uint256, Foo):
-    return 1, Foo(2)
-    """,
-    """
-a: HashMap[uint256, uint256]
-
-@external
-def foo() -> HashMap[uint256, uint256]:
-    return self.a
-    """,
-    """
-event Foo:
-    a: uint256
-
-@external
-def foo(x: Foo):
-    pass
-    """,
-    """
-@external
-def foo(x: HashMap[uint256, uint256]):
-    pass
-    """,
-    """
-@external
-def foo(x: (uint256, uint256)):
-    pass
-    """,
-    """
-event Foo:
-    a: uint256
-
-foo: Foo
-    """,
-    """
-event Foo:
-    a: uint256
-
-@external
-def foo():
-    f: Foo = Foo(1)
-    pass
-    """,
-    """
-event Foo:
-    a: uint256
-
-b: HashMap[uint256, Foo]
-    """,
-    """
-event Foo:
-    a: uint256
-
-b: HashMap[Foo, uint256]
-    """,
-    """
-b: immutable(HashMap[uint256, uint256])
-
-@external
-def __init__():
-    b = empty(HashMap[uint256, uint256])
-    """,
-    """
-b: immutable((uint256, uint256))
-
-@external
-def __init__():
-    b = (0, 0)
-    """,
-    """
-@internal
-def _foo() -> (uint256, uint256):
-    return 1, 2
-
-@external
-def foo():
-    a: (uint256, uint256) = self._foo()
-    """,
 ]
